update svg and remove shit
